applications/treconomics_project3/requirements.txt
beautifulsoup4==4.4.0
breadability==0.1.20
chardet==2.3.0
cssselect==0.9.1
datrie==0.7
diff-match-patch==20121119
Django==1.7.8
django-import-export==0.2.8
docopt==0.6.2
feedparser==5.1.3
jieba==0.37
lxml==3.4.4
nltk==3.0.4
numpy==1.9.2
Pillow==2.9.0
pyteaser==2.0
pytz==2015.4
redis==2.10.3
selenium==2.47.0
tablib==0.10.0
Whoosh==2.7.0
